k/alter_partitions: add check for undefined topic
One of the validators in the AlterPArtitionReassiignments API uses topic
metadata to check the topic replication factor. However, it is possible
that the request went to a follower and that node may have undefined
metadata because the node did not get updates yet. Therefore, this
commit adds a check for undefined topic metadata.