cloud_storage: use fragmented_vector in directory walk
Use of a plain std::vector is unsafe if very large numbers
of files are in the cache.

It would be better to improve this code to not require
a monolithic vector of files at all, but this change mitigates
it by avoiding the vector risking a bad_alloc under memory
fragmentation.