k/quotas: metrics: traffic instead of throughput
The 'throughput' gauge is removed because it did not reflect actual TP.
That was the distance of current level of tokens from the maximum level
(how empty the bucket is), divided by token bucket window width. However
that value has nothing to do with actual TP, it says something about
the size of the chunk received via Produce, but not about the TP. And
indeed, the figure for that was bigger the larger message size in the
test was. Actually, there is no measure of actual TP in a token bucket
alone: tocket bucket is designed to limit TP by comparing it to a model
TP rather than by measuring it.

A new counter metric added: 'traffic'. This is the number of bytes
received/sent by the shard; customers will be able to differentiate
that value in Grafana to get the TP number. The counter will overflow
at 18 exabytes; even when it does, it won't cause any disruption to
the result of non-negative derivative that should be used to get TP
from it.

Ingress 'traffic' counter reflects the bytes taken into processing,
and does not reflect the bytes that are still waiting in e.g.
forced throttling stage. This way, this counter is suitable for measuring
the throttled TP.