cloud_storage,cluster: Restrict status for recovery service
The topic recovery service used to store the status of the recovery on
all shards, even though the recovery only ever runs on shard 0. The
reasoning behind this was that any shard could tell if the recovery was
active and so abort a duplicate recovery request. However this causes a
problem with extended status, where a map of download counts and the
request used for recovery also needs to be copied to each shard.

Additionally the admin server now restricts requests to initiate the
recovery process to controller leader, so it is not required to copy the
status to each shard. This change restricts the status of recovery to be
stored only on shard 0. The requests to check the status are also
invoked on shard 0.