admin_server: Enable returning extended recovery status
The recovery status has been extended to contain multiple entries. This
change adds an optional query param called "extended". This param, when
passed in as true or 1, returns a list of the past five statuses of the
recovery service.

The default behavior is unchanged and returns the latest status of the
service.