rpk: trim controllers logs in debug bundle
If the total size of the logs exceeds the
specified limit, we will reduce the size of the
controller log directory. Specifically, we will
keep the first and last 'limit/2' bytes of the
log files, discarding the middle section to bring
the total size of the logs under the limit.