tests: run unit tests with explicit core count
Otherwise, we end up running 48 threads on 48 vCPU nodes,
which makes the tests slower, and also makes execution less
deterministic when switching from environments.

Any development machine has at least 8 cores, so pick
this as an arbitrary "many cores" count for the tests
that do not explicitly run on a single core.