scheduling: made constraints more flexible
Made interface of `hard_constraint` and `soft_constraint` more flexible.
Previously an instance of a constraint evaluator was immutable and were
not modified during the whole allocation process this made some
constraints tightly coupled with the allocation process as they had to
be aware of intermediate allocations. f.e. rack awareness constraint.

Change constraint interface to be a factory for constraint evaluator.
Each time an allocation is required a constraint interface
`make_evaluator` method is called with current replica set. This way a
factory is able to make some preliminary calculations before evaluating
constraint for the possible set of options. Additionally constraints are
now decoupled from an intermediate allocation making it possible to
create them outside of `partition_allocator` itself, even if they relay
on intermediate allocations.

This implementation will make it possible to define all constraint types
per partition instead of relaying on partition_allocator to decide which
constraints to use.

Signed-off-by: Michal Maslanka <michal@redpanda.com>