s/disk_log_impl: return "Raft start timestamp"
Previously, `disk_log_impl::start()` returned the base timestamp of the
the first segment presetn in the log. However, that segment may have
been marked for eviction and, therefore, all offsets within fall below
the Raft start offset.

This function is used in the timequery logic in order to prevent
returning an offset that's below the Raft start offset. The old
semantics made us return offsets below the Raft start offset instead,
which was a bug.