consensus: remove start offset check in timequery
Previously, `consensus::timequery()` used to check if the offset which
timequery picked is below the Raft start offset. If that was the case,
the result would be overwritten by the Raft start offset. This makes
sense in a non tiered storage world (minus the overwriting), but if
tiered storage is enabled we should query the remote log.

The logic for this is lifted in `partition::timequery()` in a following
commit.