cluster: fix sometimes missing node version updates
Calculating deltas on health reports was not robust,
because after a leadership change theq controller can come
up on a node that already has a health report showing
a newer version, i.e. no delta wrt the next report.

Instead, generate an update to _node_versions whenever
we see a version that is newer than that in the map, and
clear _node_versions on leadership/term changes.  This
guarantees that within each controller term, we will
pay attention to node versions from all nodes until we
have accumulated a version from each node, and thereafter
we will only submit updates if the version in a health
report is newer than the one in _node_versions.

Fixes: https://github.com/redpanda-data/redpanda/issues/8758