tests: cache release list downloaded from github
Github rate limits anonymous HTTP clients: if one runs
many tests in parallel, this can result in 403s.

We may assume that the set of released packages does not change
within the lifetime of a ducktape process, so cache this
as a class member.

Populating the cache is protected by a mutex so that when N
tests start concurrently, they do not all enter the code for
populating the cache.