cloud_storage: downgrade DeleteObject 404 message to debug
This message is frequent on GCS backends: INFO is too verbose,
and the message does not indicate an issue that users should pay
attention to.

This is a bespoke change for v22.3.x, because on `dev` the code
has already been revised in a way that changed this along with
more invasive changes.