tests: wait for listobjects to work after bucket creation
This is a workaround for an apparent quirk in AWS S3,
where ListObjectsv2 requests returned NoSuchBucket
errors when run soon after creating a bucket.

I didn't really want to put this hack in, but the most recent
debug logs point pretty unambiguously to a lack of read-after-write
consistency in the API when it comes to listobjectsv2.

Fixes: https://github.com/redpanda-data/redpanda/issues/8490
Fixes: https://github.com/redpanda-data/redpanda/issues/8564
Fixes: https://github.com/redpanda-data/redpanda/issues/8754