cloud_storage: retry if tx manifest is not present
Previously, hydration was only retried if the segment was removed from
the cache mid-way through. Turns out the fetching of the tx manifest
races with the cache purging:
1. `run_hydrate_bg` is called with both the segment and tx manifest
   present in the cache
2. We fall through and materialise the segment in `do_hydrate_segment`.
3. The segment and tx manifest are evicted from the cloud storage cache
4. We attempt to materialise the tx manifest in `do_hydrate_txrange`.
   This fails, and `false` is returned to signal a retry.
5. Since `_data_file` has been set at step (2), nothing is actually
   retried. This way, we end up with a remote segment that has not
   materialised it's aborted transactions.

The fix itself is very simple: ensure that we don't exit
`run_hydrate_bg` until both the segment and tx manifest have been
materialised.