test: upgrade arroyo dependency
Our tests were running with arroyo 1.0.0, which is fairly old compared
to the latest release of 2.5.0 to which this PR upgrades the dependency.

The assertion failures in #6826 were removed in arroyo 2.3 with this
commit https://github.com/getsentry/arroyo/commit/84934df1e14f1490682adf91b208e004b23bf82e.

Installing the latest arroyo also bumps the version of the
confluent-kafka-python library, which includes a change which requires a
deprecation flag to be set. See here for more details:

https://github.com/confluentinc/confluent-kafka-python/blob/17029ada91e288842a10e2e4886638fb2de2b4f0/CHANGELOG.md?plain=1#L115

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>