archival: refine archiver pause logic
Previously we would drop out between writing segments
and writing manifest if we were transferring leadership.

What we want during a transfer is for the node to finish its
segment uploads, update the archival_metadata_stm so that
the new leader does not forget about these uploaded segments,
and upload the manifest to S3 so that the written segment
becomes visible promptly without having to wait for the
new leader to do its next segment upload to write the manifest.