rptest: improve timequery below start offset test
Previously, `test_timequery_after_segments_eviction` expected that
timequeries for offsets within a segment would be serviced up until the
segment was removed from disk. This assumption is not correct, as we
should not serve timequeries that fall below the start offset of the
local log (unless cloud storage is enabled and the offset is in the
cloud log).

This commit removes the mentioned test, and adds a new test that checks
that local timequeries don't return data below the start offset of the
local log.