rptest: improve timequery below start offset test
CONFLICT:
- This patch was already mostly merged as a part of c41022f, but missed
  out on this bit to update segment size for non-cloud-storage clusters,
  and a removal of a test. This meant the log would not roll a segment,
  and therefore never have anything viable for GC, hence the failure to
  advance the start offset.

Previously, `test_timequery_after_segments_eviction` expected that
timequeries for offsets within a segment would be serviced up until the
segment was removed from disk. This assumption is not correct, as we
should not serve timequeries that fall below the start offset of the
local log (unless cloud storage is enabled and the offset is in the
cloud log).

This commit removes the mentioned test, and adds a new test that checks
that local timequeries don't return data below the start offset of the
local log.

(cherry picked from commit 06c1a2c6b8a5b18d8fa369a93e9d1501398f9555)