Use topic_table to check deletion in fuzzy_test
Prev validation for topic deletion uses api
that check topic in partition_manager, but
it is only check does node contains this
topic or not.

But if node read controller log not super fast
it can contains expired data in metadata_cache

So in our case slow node can contains info about
deleted topic. So in validation we should use
another way to check