tests: add CloudStorageType.AUTO
This is for tests such as scale tests, that expect to run
in an environment where only one type of object storage
is available (i.e. on EC2 nodes to use S3, on Azure nodes to use ABS).

When we run these tests in docker environment it is just for hacking
on the test, so it makes sense to pick an arbitrary backend rather
than parametrizing on ABS+S3.