archival: Do not instanciate merging job for compacted
topics. The adjacent_segment_merger_job is designed to not to upload
compacted seegments. But when the job is running on compacted partition
it can still reupload segments which are not compacted yet. It's not
necessary because compacted segment reupload will eventually reupload
partition data after compaction will happen. This may also cause CI
failures because adjacent segment merging may prevent compacted segment
reupload from happening because it will aggregate the segments in the
cloud and the compacted segment reupload has to be aligned with that
large non-compacted segment. This changes the timings of operations
(more segments has to be compacted) and the test fails.