cluster: include under replicated replicas in health
This is an additional overhead in the health monitoring
system, but it is highly valuable information: under
replicated partitions are a key signal on whether it
is safe to proceed with maintenance.

Also update the get_cluster_health_overview logic to
cap the number of partitions we will try to include
in response: this was previously a problem on systems
with >10k partitions that would try and include all
of them.  Visually bad in the CLI, and a risk of bad_alloc.