admin: add partitions/local_summary API
This is a way to fetch high level stats for the partition replicas
on the node, without going via the health monitor's central view.

The API is explicitly called 'local_summary' rather than 'summary'
to avoid the risk of confusion from developers thinking that this
API refers to all partitions in the cluster, rather than just those
partitions with a replica on this node.

This may be useful for tests or automation that would like to
directly read the leaderless+under-replicated counts from
all nodes, to avoid any risk of missing issues due to delays
in updates to the central health state.