archival: Double check jobs_available
There is a scheduling event that occurs between the predicate in
`ss::condition_variable::wait` returning true and the `wait` method
returning, meaning that upon return from that method, there is no
guarantee that the predicate is true anymore.

Adding an additional check for `jobs_available()` to avoid the
`vassert`.

Fixes: #8964
Signed-off-by: Michael Boquard <michael@redpanda.com>