tests: retry AlterPartitions on topic unknown
AlterPartitions will return unknown_topic_or_partition when a check for
the topic's metadata fails. This error is likely because of stale topic
metadata and a simple retry will resolve the issue. Therefore, this
commit adds a new unit test and it adds retry loops to ducktape tests.

Fixes #8824
Fixes #8825