tests: set trace log level on shutdown timeout
This updates the shutdown timeout hook to set the log level to trace and
wait for 10 seconds. The idea here is that if there is any funny
business (e.g. something causing a hang), it will be more visible with
trace logging enabled, especially in cases where info logs are the
default (e.g. nightly CDT runs).