features: don't set original version on fast-forward
The original version should reflect the first cluster-wide
activated version, not the value of earliest_version when
a node started.

This change enables the feature table to have an active
version without having an original_version: this is a sign
that the active version is something set ephemerally during
startup on a fresh node, rather than something set persisently
via the controller log, or loaded from a snapshot.