cloud_storage: azure shared key live reload
Previously, updating the shared key used for Azure auth required a
cluster restart. This is incovenient for customers that wish to rotate
keys.

This commit makes the shared key reload without a restart by listening
for changes on the configuration property and rebuilding the static
credentials accordingly. If the property becomes unset (likely a user
error), we keep using the current key until the cluster is re-started.