tests: rm err handling when no active reassignment
This test checks for successfully canceling an on-going partition
reassignment. Previously, the test would catch errors when there is no
reassignment in progress. This is a problem because it covers-up
potential failures. Instead, remove error handling and increase the
amount of data in the topic before executing the partition move.