tests: update transactions check_parsing_test
This test used latest version environment variable override
to impersonate an older Redpanda.  Because the impersonated version
was older than our current binary's earliest version, this
wasn't working: we had to override earliest version as well.

We should consider removing this test or modifying it to use
real old binaries rather than relying on the current binary
to be able to authentically output structures in the format
of logical version 5 for upgrade testing.