config: fix binding assignment operator
The assignment operator assumes that the lhs is not an initialized
binding. When this assumption is violated the lhs might already be on a
property's binding list, which can result in the binding be added
to a new list without being removed from its current list. In release
mode it seems that the assertions in the boost auto-unlink hooks
are compiled out, leading to mysterious segfaults.

The `_abort_on_oom` binding in application.cc::initialize is such a
binding. The re-assignment occurs in the kafka_server and kafka_client
tests which "reset" the application and call initialize multiple times.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>