test: skip cloud storage upgrade test in Azure CDT
By default, the RedpandaService start-up logic is going to set the
correct cluster configurations for the cloud the test is running in.
This means that, it will set the azure cloud storage cluster configs
when running in Azure. This is problematic for upgrade tests that start
from a version which predates Azure.

This commit skips all cloud storage upgrade tests when running in Azure
CDT by moving the redpanda service start-up in the body of the test and
using a new skip decorator.