iobuf: fix empty fragment byte iteration
The iobuf byte iterator assumed that an iobuf didn't contain any empty
fragments (size, not capacity). This meant that when initialized or when
moving to the next fragment that it would dereference the first byte in
the fragment even if it were not user data. The patch fixes this by
skipping fragments when they are empty.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>