Move schema util implementation to header
We have a clientutils library which contains a single source file:

cluster_config_schema_util.cc

which itself only has a couple dozen lines of code. This code is
shared between the redpanda binary and rp_util, hence it's
separate existence in clientutils library.

As part of a cleanup of the repdanda module, it makes sense
to move this small implementation to a header file: it is only
included by 2 .cc files in the whole code base and this removes a
library from the build and avoids some questions about
where it should live (since it should not really be hiding out
in the redpanda module, etc).