cloud_storage: track cloud log size in manifest
This commit introduces a new field to the prologue of the partition
manifest. 'cloud_log_size_bytes' tracks the size of the cloud log (i.e.
sum of the sizes of uploaded log segments).

All operations on the cloud log are supported: addition, truncation,
compaction and merging. Truncation, compaction and merging decrease the
size of the cloud log. This decrease is reflected in the manifest
*before* the files have been deleted from cloud storage.

In order to understand this decision, consider the following:
* The cloud log size will be used in our billing infra, so it's vital to
  avoid over-counting.
* Currently, the cloud storage housekeeping loop updates the partition
  manifest (via the STM) only after all the desired deletes succeeded.

Assuming that all deletions have succeeded apart from one, and we get
queried by the billing service at the wrong time, we would over-report
the amount of bytes in S3.