rpk: fix panic when listing returns empty containers
When we have a running cluster we try to inspect
the running containers and use some properties
from the client response (IP, status, etc...).

However, some clients (e.g Podman) might return
an empty container object and that produced a
panic in rpk because we were not properly
handling the objects.