tests: improve ArchivalTest.test_isolate
This test simply asserted on key count: that fails if
the server had uploaded an empty manifest for partitions.

Instead, use an S3Snapshot to check what we specifically
care about (that there are no data segments uploaded), and
also add a control at the end of the test that we do see
data segments when isolation ends, both as segment objects
and in the manifests.