tests: don't set {assert} logs on shutdown
The {assert} logger space is used for vassert() calls and only logs
ERRORs. It's not useful to set it to TRACE, and can result in alarming
misdiagnoses as node crashes.

I considered adding a allowed log check for when we use the admin server
to set {assert} logs, but this seems like a nicer approach in that it
avoids the "bad" logs lines altogether.