cluster: report logical versions in metrics_reporter
This is useful for knowing the "logical age" of a cluster
to understand how much of the deployed footprint is newer
clusters vs. clusters that have lived through upgrades.

Reporting the per-node logical version also enables detecting
mid-upgrade systems (or detecting systems which have been
incorrectly left in a mixed-version state)