tests: rework S3Snapshot into BucketView
For tests that would like to inspect manifests,
the approach of always doing a bucket listing doesn't
scale well for any tests that use larger volumes of data.

We have enough information to calculate manifest paths directly,
once we fetch revision IDs from the admin API.  So we can do that
for manifest path queries, and only bother doing the bucket listing
if someone tries to fetch one of the class members that actually
requires it.

The name is changed to reflect that we no longer only support S3,
and that the class is more of a caching view than a snapshot.