archival: wait for upgrade before manifest uploads
Previous commits have introduced a new field to the prologue of the
partition manifest. This means that previous Redpanda versions, will not
be able to decode the manifest, which can happen if a recovery is
triggered during the upgrade and in the case of read replicas.

This commit introduces a feature flag that stalls partition manifest
updates until the entire cluster has been upgraded. Note that this does
not help with the read replica case. The solution there is to upgrade
read replica clusters first.

Note that the feature flag was intentionally named in a very generic
way. The next release, v23.2, will change the manifest format entirely
and the intention is to use this flag to guard that change.