configuration: Add cloud storage backend
A new property "cloud_storage_backend" defined by an enum is added to
the cluster config. This property is optional and if not specified will
be inferred from other configuration parameters such as the api
endpoint. This proprty will be used to drive API calls which depend on
backend capabilities. For example Google storage does not support
multi-object delete, so we should delete objects in sequence if running
on a google backend. However AWS and Minio do support this, so there we
should delete objects in batches.