storage: remove dubious-looking dead code
Without knowing the `is_collectible()` condition and the "unlikely"
condition are mutually exclusive, it looks as though we're abandoning a
futurized `gc()`, which is quite alarming.

From John:
"Then 95320bc6087 fixed the bug by adding the is_collectable check, and
the unlikely block has been dead code ever since."

Note: I had to tweak this commit in order to backport to v22.3.x since
the name of the gate involved changed from `_compaction_gate` to
`_compaction_housekeeping_gate` in dev. I've used the old name in this
commit.