utils/delta_for.h: constexpr of delta_* algos to use them as nttp
this in done in a retro-compatible way, so as to not touch preexisting run-time use.
delta_for_encoder can now be used with a runtime version of delta_* algo
or with a nttp value.
at the same time, delta_for_encoder and delta_stream_pos_t are
serde-enabled, and serde is aware if delta_* is nttp or not, reducing
slightly the amount of data used