cloud_storage/segment_meta_cstore: serde conversion and unit test
serde is done naively via serde::envelope. std::list and absl::btree_map
are not serde enabled, they are processed in their containing class via
serde_write/read mirroring what it's done for other sequence/map types
(size,[elements...]/[(key,value)...]).
the nttp version of delta_* algorithm is used to remove the need to
serde it. this has a cascading effect of removing the need for some user
defined constructors.
column_store::hint_vec_t is transformed from a std::tuple to a
std::array since all the elements of the tuple were of type hint_t.
memory and performance wise there should be no difference, but it's
compatible with serde.

a unit test is added that serializes and deserializes a
segment_meta_cstore and checks the elements