cluster: expose cloud storage size on partition
This commit adds a method to `cluster::partition` that exposes a new
`cloud_log_size` method. It returns the sum of all *tracked* log
segments that have beeb uploaded to cloud storage.

A segment is tracked if it is present in the manifest and has not been
slated for removal.

(cherry picked from commit bdab9f23c8687b8ab0966b320bce0727ffcb6860)