redpanda/tests: add fixture mode for archiver tests
These tests used to instantiate a partition with
tiered storage enabled, and then stop+replace the
archiver.

That is awkward, and causes problems now that the archiver
does an initial upload of an empty manifest before
any data is written.

Instead, run the test with tiered storage config in
place, but not _enabled_, so that when the test
constructs an archiver, that's the first time it
is constructed.