cluster: include initial revision in leader metadata
When propagating leadership information Redpanda include a partition
revision id to fence the updates coming from older instance of a
partition with the same ntp. The revision that is used in the leader
table is increased every time the partition configuration change. Added
an initial revision id to the leadership metadata. Initial revision
changes when partition is recreated but not when it is moved.

Signed-off-by: Michal Maslanka <michal@redpanda.com>