tests: remove UpgradeFrom22_2_7VerifyMigratedRetentionSettings
This test covers the hop from 22.2 to 22.3 specifically: we
do not need to carry the test indefinitely.

We could retain this test by updating it to step through all
intermediate versions, but it's not worth it: when we upgrade
to latest redpanda, the system will have passed through the
already-tested 22.3 code that handles the upgrade.