rpk/debug bundle: use fqdn in admin API addresses
We were using IPs but that's not suitable for
brokers with TLS enabled since the IP won't be
in the certificate.

Instead we are going to use the FQDN to talk to
the rest of the cluster's admin API addresses.

(cherry picked from commit d960a5ab847490b259bf24c03f7a44140f31abb6)