tests: ability to dynamically parametrize cloud_storage_type
Replaced `parametrized` with `matrix` decorator,
and the cloud_storage_type holds the parametrization
needed for each host type. For example for
cdt runs on aws, the matrix will only contain s3
as cloud storage type. same for gcp or azure.
For regular tests (host=docker container), it
parametrizes with both s3 and abs

ref redpanda-date/redpanda#8704