tests: do not use a seed server in folder deletion test
In folder deletion test we expect a node id change as the node with old
id is going to be decommissioned. To ensure the node id change we must
not use a seed server as this is a cluster founding broker for which
node id doesn't change.

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 2788ac9afb114b66c1a6ce3beaee6d0e55a60b13)