test: update ShadowIndexingCacheSpaceLeakTest
This test was setting total retention to 1 segment,
and relying on the laziness of retention enforcement
for any reads to hit S3.  This was probably written
before `retention.local.target.bytes` was a thing.

Update the test to specify separate local + remote
retention settings, so that the test's expectation
of reads reliably hitting remote storage will be met.