cloud_storage: Add segment_meta v3
The v3 is the same as v2 but metadata_size_hint in the segment_meta
struct is used to indicate that tx-manifest is uploaded to S3. In the
short term this is needed to avoid doing extra request to S3. In the
long run this field will be used in v4 to indicate the size of the
metadata embedded into a segment object.