c/topic_table: fix deallocation with canceled updates.
If an update is canceled, the current assignment in topic table
state is reset to 'previous_replicas' (effectively switching the
direction of the move that is being canceled) but a racing delete that
computes in progress assignments for deallocation doesn't factor this
in.

This was resulting in fewer than actual deallocations thus blocking
decommission. Additionally adds some logging for debugging similar
issues in the future.