cloud_storage: Split initialization of 'remote'
Currently, cloud_storage::remomte contains
cloud_storage_clients::client_pool instance. It's created in the c-tor
individually on every shard. In order to be able to move connections
between shards the client_pool should be a peering_service. To make this
possible the client_pool has to be initialized in the application and
passed to the c-tor of the cloud_storage::remote.

As a consequence a lot of tests has to be changed. In most cases the
cloud_storage::remote was turned into a sharded<remote> and initialized
in the fixture.